+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Security in case of vulnerabilities.
 
 ## [Unreleased]
 
+
+## [25.11.2] - 2025-11-02
+
 ### Added
 - Support for excluding archived repositories
 
@@ -251,7 +254,8 @@ Security in case of vulnerabilities.
 - First version suitable for public consumption.
 
 
-[Unreleased]: https://github.com/nkantar/Starminder/compare/25.10.20.2...HEAD
+[Unreleased]: https://github.com/nkantar/Starminder/compare/25.11.2...HEAD
+[25.11.2]: https://github.com/nkantar/Starminder/compare/25.10.20.2...25.11.2
 [25.10.20.2]: https://github.com/nkantar/Starminder/compare/25.10.20.1...25.10.20.2
 [25.10.20.1]: https://github.com/nkantar/Starminder/compare/25.10.20...25.10.20.1
 [25.10.20]: https://github.com/nkantar/Starminder/compare/25.10.19...25.10.20